backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Devin AI <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.116.1"}
psycopg = {extras = ["binary"], version = "^3.2.9"}
pillow = "^11.3.0"
pytesseract = "^0.3.13"
python-multipart = "^0.0.20"
groq = "^0.30.0"
python-dotenv = "^1.1.1"
pyjwt = "^2.8.0"
cryptography = "^42.0.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
reportlab = "^4.0.0"
PyPDF2 = "^3.0.1"
PyMuPDF = "^1.23.0"
pdf2image = "^1.16.3"
qrcode = {extras = ["pil"], version = "^8.2"}
aiosmtplib = "^3.0.0"
email-validator = "^2.1.0"
apscheduler = "^3.10.4"
supabase = "^2.2.0"
asyncpg = "^0.29.0"
bleach = "^6.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
